correção layout e filtros

+++ app/Http/Controllers/ExportController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Inertia\Inertia;
 use App\Models\Export;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
@@ -10,14 +11,29 @@ class ExportController extends Controller
 {
     public function index()
     {
-        $exports = Export::paginate(15);
+        $exports = Export::paginate(5);
+
+        return Inertia::render('Reports', [
+            'exports' => $exports
+        ]);
+    }
+
+    public function show($export) {
+        $export = Export::find($export);
+
+        return Storage::download($export->file_name);
     }
 
     public function destroy(Export $export)
     {
-        Storage::delete($export->filename);
-        $export->delete();
+        $export = Export::find($export);
 
-        return "Deletado";
+        if ($export) {
+            Storage::delete($export->file_name);
+            $export->delete();
+        }
+
+        return redirect()->back()
+            ->with('success', 'Seu arquivo foi removido com sucesso');
     }
 }
